Misc Bugfixes (#1015)
* Fixed some security issues in dev env

* When deleting folders in bookmark cleanup, delete empty folders correctly.

* When a new library is created and cards are added, cards can have a blank library name. Card library name code is reworked to be much lighter on memory.

* Added a config for github issues to disable blank issues.

* Skip any sort of directory iteration code if we haven't deleted any bookmarks.

* Fixed a bug where some style overrides were duplicating. Now logic is much more targetted, only applying to the correct tags.

* Applied sorting to the filtering apis.

* Reverted one of my changes for a better version Robbie did.

+++ API/Services/Tasks/CleanupService.cs
@@ -184,19 +184,23 @@ namespace API.Services.Tasks
 
 
             var filesToDelete = allBookmarkFiles.ToList().Except(bookmarks).ToList();
-            _logger.LogDebug("[Bookmarks] Bookmark cleanup wants to delete {Count} files", filesToDelete.Count());
+            _logger.LogDebug("[Bookmarks] Bookmark cleanup wants to delete {Count} files", filesToDelete.Count);
+
+            if (filesToDelete.Count == 0) return;
 
             _directoryService.DeleteFiles(filesToDelete);
 
             // Clear all empty directories
-            foreach (var directory in _directoryService.FileSystem.Directory.GetDirectories(bookmarkDirectory))
+            foreach (var directory in _directoryService.FileSystem.Directory.GetDirectories(bookmarkDirectory, "", SearchOption.AllDirectories))
             {
-                if (_directoryService.FileSystem.Directory.GetFiles(directory).Length == 0 &&
+                if (_directoryService.FileSystem.Directory.GetFiles(directory, "", SearchOption.AllDirectories).Length == 0 &&
                     _directoryService.FileSystem.Directory.GetDirectories(directory).Length == 0)
                 {
                     _directoryService.FileSystem.Directory.Delete(directory, false);
                 }
             }
+
+
         }
     }
 }
